requirements.txt
# requirements.txt

# Core dependencies for both ingestion scripts and backend services

fastapi>=0.95.0  # FastAPI web framework
uvicorn[standard]>=0.22.0  # ASGI server for FastAPI
beautifulsoup4>=4.12.2  # HTML parsing
readability-lxml>=0.8.1  # Main article extraction
requests>=2.31.0  # HTTP requests
chromadb==0.4.15  # Vector store
sentence-transformers>=2.2.2   # Local embedding model
python-dotenv>=1.0.0  # Environment variable management